+++ test/cpp/jit/test_constant_pooling.cpp
@@ -1,9 +1,10 @@
+#include <gtest/gtest.h>
+
 #include <torch/csrc/jit/ir/ir.h>
 #include <torch/csrc/jit/ir/irparser.h>
 #include <torch/csrc/jit/passes/constant_pooling.h>
 #include <torch/csrc/jit/passes/constant_propagation.h>
 #include <torch/csrc/jit/testing/file_check.h>
-#include "test/cpp/jit/test_base.h"
 
 #include <sstream>
 #include <string>
@@ -11,26 +12,26 @@
 namespace torch {
 namespace jit {
 
-void testConstantPooling() {
-  {
-    auto graph = std::make_shared<Graph>();
-    parseIR(
-        R"IR(
+TEST(ConstantPoolingTest, Int) {
+  auto graph = std::make_shared<Graph>();
+  parseIR(
+      R"IR(
 graph():
   %8 : int = prim::Constant[value=1]()
   %10 : int = prim::Constant[value=1]()
   return (%8, %10)
   )IR",
-        &*graph);
-    ConstantPooling(graph);
-    testing::FileCheck()
-        .check_count("prim::Constant", 1, /*exactly*/ true)
-        ->run(*graph);
-  }
-  {
-    auto graph = std::make_shared<Graph>();
-    parseIR(
-        R"IR(
+      &*graph);
+  ConstantPooling(graph);
+  testing::FileCheck()
+      .check_count("prim::Constant", 1, /*exactly*/ true)
+      ->run(*graph);
+}
+
+TEST(ConstantPoolingTest, PoolingAcrossBlocks) {
+  auto graph = std::make_shared<Graph>();
+  parseIR(
+      R"IR(
 graph(%cond : Tensor):
   %a : str = prim::Constant[value="bcd"]()
   %3 : bool = aten::Bool(%cond)
@@ -44,17 +45,18 @@ graph(%cond : Tensor):
   %7 : (str, str) = prim::TupleConstruct(%a, %b)
   return (%7)
   )IR",
-        &*graph);
-    ConstantPooling(graph);
-    testing::FileCheck()
-        .check_count("prim::Constant[value=\"abc\"]", 1, /*exactly*/ true)
-        ->check_count("prim::Constant[value=\"bcd\"]", 1, /*exactly*/ true)
-        ->run(*graph);
-  }
-  {
-    auto graph = std::make_shared<Graph>();
-    parseIR(
-        R"IR(
+      &*graph);
+  ConstantPooling(graph);
+  testing::FileCheck()
+      .check_count("prim::Constant[value=\"abc\"]", 1, /*exactly*/ true)
+      ->check_count("prim::Constant[value=\"bcd\"]", 1, /*exactly*/ true)
+      ->run(*graph);
+}
+
+TEST(ConstantPoolingTest, PoolingDifferentDevices) {
+  auto graph = std::make_shared<Graph>();
+  parseIR(
+      R"IR(
 graph():
   %2 : int = prim::Constant[value=2]()
   %1 : int = prim::Constant[value=1]()
@@ -70,22 +72,21 @@ graph():
   prim::Print(%x, %y, %z)
   return (%1)
   )IR",
-        &*graph);
-    // three tensors created - two different devices among the three
-    // don't have good support for parsing tensor constants
-    ConstantPropagation(graph);
-    ConstantPooling(graph);
-    testing::FileCheck()
-        .check_count(
-            "Float(2:1, requires_grad=0, device=cpu) = prim::Constant",
-            1,
-            /*exactly*/ true)
-        ->check_count(
-            "Long(2:1, requires_grad=0, device=cpu) = prim::Constant",
-            1,
-            /*exactly*/ true)
-        ->run(*graph);
-  }
+      &*graph);
+  // three tensors created - two different devices among the three
+  // don't have good support for parsing tensor constants
+  ConstantPropagation(graph);
+  ConstantPooling(graph);
+  testing::FileCheck()
+      .check_count(
+          "Float(2:1, requires_grad=0, device=cpu) = prim::Constant",
+          1,
+          /*exactly*/ true)
+      ->check_count(
+          "Long(2:1, requires_grad=0, device=cpu) = prim::Constant",
+          1,
+          /*exactly*/ true)
+      ->run(*graph);
 }
 } // namespace jit
 } // namespace torch
