+++ test/cpp/jit/test_create_autodiff_subgraphs.cpp
@@ -1,4 +1,5 @@
-#include "test/cpp/jit/test_base.h"
+#include <gtest/gtest.h>
+
 #include "test/cpp/jit/test_utils.h"
 
 #include "torch/csrc/jit/passes/create_autodiff_subgraphs.h"
@@ -6,7 +7,7 @@
 namespace torch {
 namespace jit {
 
-void testCreateAutodiffSubgraphs() {
+TEST(CreateAutodiffSubgraphsTest, Basic) {
   auto graph = build_lstm();
   CreateAutodiffSubgraphs(graph, /*threshold=*/2);
   // all of the ops are within the DifferentiableGraph
