+++ test/cpp/jit/test_custom_operators.cpp
@@ -1,4 +1,5 @@
-#include "test/cpp/jit/test_base.h"
+#include <gtest/gtest.h>
+
 #include "test/cpp/jit/test_utils.h"
 
 #include "torch/csrc/jit/ir/alias_analysis.h"
@@ -11,134 +12,135 @@
 namespace torch {
 namespace jit {
 
-void testCustomOperators() {
-  {
-    torch::RegisterOperators reg(
-        "foo::bar", [](double a, at::Tensor b) { return a + b; });
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::bar"));
-    ASSERT_EQ(ops.size(), 1);
+TEST(CustomOperatorTest, InferredSchema) {
+  torch::RegisterOperators reg(
+      "foo::bar", [](double a, at::Tensor b) { return a + b; });
+  auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::bar"));
+  ASSERT_EQ(ops.size(), 1);
 
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::bar");
+  auto& op = ops.front();
+  ASSERT_EQ(op->schema().name(), "foo::bar");
 
-    ASSERT_EQ(op->schema().arguments().size(), 2);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "_0");
-    ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
-    ASSERT_EQ(op->schema().arguments()[1].name(), "_1");
-    ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
+  ASSERT_EQ(op->schema().arguments().size(), 2);
+  ASSERT_EQ(op->schema().arguments()[0].name(), "_0");
+  ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
+  ASSERT_EQ(op->schema().arguments()[1].name(), "_1");
+  ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
 
-    ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
+  ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
 
-    Stack stack;
-    push(stack, 2.0f, at::ones(5));
-    op->getOperation()(&stack);
-    at::Tensor output;
-    pop(stack, output);
+  Stack stack;
+  push(stack, 2.0f, at::ones(5));
+  op->getOperation()(&stack);
+  at::Tensor output;
+  pop(stack, output);
 
-    ASSERT_TRUE(output.allclose(at::full(5, 3.0f)));
-  }
-  {
-    torch::RegisterOperators reg(
-        "foo::bar_with_schema(float a, Tensor b) -> Tensor",
-        [](double a, at::Tensor b) { return a + b; });
-
-    auto& ops =
-        getAllOperatorsFor(Symbol::fromQualString("foo::bar_with_schema"));
-    ASSERT_EQ(ops.size(), 1);
-
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::bar_with_schema");
-
-    ASSERT_EQ(op->schema().arguments().size(), 2);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "a");
-    ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
-    ASSERT_EQ(op->schema().arguments()[1].name(), "b");
-    ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
-
-    ASSERT_EQ(op->schema().returns().size(), 1);
-    ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
-
-    Stack stack;
-    push(stack, 2.0f, at::ones(5));
-    op->getOperation()(&stack);
-    at::Tensor output;
-    pop(stack, output);
-
-    ASSERT_TRUE(output.allclose(at::full(5, 3.0f)));
-  }
-  {
-    // Check that lists work well.
-    torch::RegisterOperators reg(
-        "foo::lists(int[] ints, float[] floats, Tensor[] tensors) -> float[]",
-        [](torch::List<int64_t> ints,
-           torch::List<double> floats,
-           torch::List<at::Tensor> tensors) { return floats; });
-
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::lists"));
-    ASSERT_EQ(ops.size(), 1);
-
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::lists");
-
-    ASSERT_EQ(op->schema().arguments().size(), 3);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "ints");
-    ASSERT_TRUE(
-        op->schema().arguments()[0].type()->isSubtypeOf(ListType::ofInts()));
-    ASSERT_EQ(op->schema().arguments()[1].name(), "floats");
-    ASSERT_TRUE(
-        op->schema().arguments()[1].type()->isSubtypeOf(ListType::ofFloats()));
-    ASSERT_EQ(op->schema().arguments()[2].name(), "tensors");
-    ASSERT_TRUE(
-        op->schema().arguments()[2].type()->isSubtypeOf(ListType::ofTensors()));
-
-    ASSERT_EQ(op->schema().returns().size(), 1);
-    ASSERT_TRUE(
-        op->schema().returns()[0].type()->isSubtypeOf(ListType::ofFloats()));
-
-    Stack stack;
-    push(stack, c10::List<int64_t>({1, 2}));
-    push(stack, c10::List<double>({1.0, 2.0}));
-    push(stack, c10::List<at::Tensor>({at::ones(5)}));
-    op->getOperation()(&stack);
-    c10::List<double> output;
-    pop(stack, output);
-
-    ASSERT_EQ(output.size(), 2);
-    ASSERT_EQ(output.get(0), 1.0);
-    ASSERT_EQ(output.get(1), 2.0);
-  }
-  {
-    torch::RegisterOperators reg(
-        "foo::lists2(Tensor[] tensors) -> Tensor[]",
-        [](torch::List<at::Tensor> tensors) { return tensors; });
-
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::lists2"));
-    ASSERT_EQ(ops.size(), 1);
-
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foo::lists2");
-
-    ASSERT_EQ(op->schema().arguments().size(), 1);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "tensors");
-    ASSERT_TRUE(
-        op->schema().arguments()[0].type()->isSubtypeOf(ListType::ofTensors()));
-
-    ASSERT_EQ(op->schema().returns().size(), 1);
-    ASSERT_TRUE(
-        op->schema().returns()[0].type()->isSubtypeOf(ListType::ofTensors()));
-
-    Stack stack;
-    push(stack, c10::List<at::Tensor>({at::ones(5)}));
-    op->getOperation()(&stack);
-    c10::List<at::Tensor> output;
-    pop(stack, output);
-
-    ASSERT_EQ(output.size(), 1);
-    ASSERT_TRUE(output.get(0).allclose(at::ones(5)));
-  }
+  ASSERT_TRUE(output.allclose(at::full(5, 3.0f)));
 }
 
-void testCustomOperatorAliasing() {
+TEST(CustomOperatorTest, ExplicitSchema) {
+  torch::RegisterOperators reg(
+      "foo::bar_with_schema(float a, Tensor b) -> Tensor",
+      [](double a, at::Tensor b) { return a + b; });
+
+  auto& ops =
+      getAllOperatorsFor(Symbol::fromQualString("foo::bar_with_schema"));
+  ASSERT_EQ(ops.size(), 1);
+
+  auto& op = ops.front();
+  ASSERT_EQ(op->schema().name(), "foo::bar_with_schema");
+
+  ASSERT_EQ(op->schema().arguments().size(), 2);
+  ASSERT_EQ(op->schema().arguments()[0].name(), "a");
+  ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
+  ASSERT_EQ(op->schema().arguments()[1].name(), "b");
+  ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
+
+  ASSERT_EQ(op->schema().returns().size(), 1);
+  ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
+
+  Stack stack;
+  push(stack, 2.0f, at::ones(5));
+  op->getOperation()(&stack);
+  at::Tensor output;
+  pop(stack, output);
+
+  ASSERT_TRUE(output.allclose(at::full(5, 3.0f)));
+}
+
+TEST(CustomOperatorTest, ListParameters) {
+  // Check that lists work well.
+  torch::RegisterOperators reg(
+      "foo::lists(int[] ints, float[] floats, Tensor[] tensors) -> float[]",
+      [](torch::List<int64_t> ints,
+         torch::List<double> floats,
+         torch::List<at::Tensor> tensors) { return floats; });
+
+  auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::lists"));
+  ASSERT_EQ(ops.size(), 1);
+
+  auto& op = ops.front();
+  ASSERT_EQ(op->schema().name(), "foo::lists");
+
+  ASSERT_EQ(op->schema().arguments().size(), 3);
+  ASSERT_EQ(op->schema().arguments()[0].name(), "ints");
+  ASSERT_TRUE(
+      op->schema().arguments()[0].type()->isSubtypeOf(ListType::ofInts()));
+  ASSERT_EQ(op->schema().arguments()[1].name(), "floats");
+  ASSERT_TRUE(
+      op->schema().arguments()[1].type()->isSubtypeOf(ListType::ofFloats()));
+  ASSERT_EQ(op->schema().arguments()[2].name(), "tensors");
+  ASSERT_TRUE(
+      op->schema().arguments()[2].type()->isSubtypeOf(ListType::ofTensors()));
+
+  ASSERT_EQ(op->schema().returns().size(), 1);
+  ASSERT_TRUE(
+      op->schema().returns()[0].type()->isSubtypeOf(ListType::ofFloats()));
+
+  Stack stack;
+  push(stack, c10::List<int64_t>({1, 2}));
+  push(stack, c10::List<double>({1.0, 2.0}));
+  push(stack, c10::List<at::Tensor>({at::ones(5)}));
+  op->getOperation()(&stack);
+  c10::List<double> output;
+  pop(stack, output);
+
+  ASSERT_EQ(output.size(), 2);
+  ASSERT_EQ(output.get(0), 1.0);
+  ASSERT_EQ(output.get(1), 2.0);
+}
+
+TEST(CustomOperatorTest, ListParameters2) {
+  torch::RegisterOperators reg(
+      "foo::lists2(Tensor[] tensors) -> Tensor[]",
+      [](torch::List<at::Tensor> tensors) { return tensors; });
+
+  auto& ops = getAllOperatorsFor(Symbol::fromQualString("foo::lists2"));
+  ASSERT_EQ(ops.size(), 1);
+
+  auto& op = ops.front();
+  ASSERT_EQ(op->schema().name(), "foo::lists2");
+
+  ASSERT_EQ(op->schema().arguments().size(), 1);
+  ASSERT_EQ(op->schema().arguments()[0].name(), "tensors");
+  ASSERT_TRUE(
+      op->schema().arguments()[0].type()->isSubtypeOf(ListType::ofTensors()));
+
+  ASSERT_EQ(op->schema().returns().size(), 1);
+  ASSERT_TRUE(
+      op->schema().returns()[0].type()->isSubtypeOf(ListType::ofTensors()));
+
+  Stack stack;
+  push(stack, c10::List<at::Tensor>({at::ones(5)}));
+  op->getOperation()(&stack);
+  c10::List<at::Tensor> output;
+  pop(stack, output);
+
+  ASSERT_EQ(output.size(), 1);
+  ASSERT_TRUE(output.get(0).allclose(at::ones(5)));
+}
+
+TEST(CustomOperatorTest, Aliasing) {
   torch::RegisterOperators reg(
       "foo::aliasing", [](at::Tensor a, at::Tensor b) -> at::Tensor {
         a.add_(b);
@@ -182,77 +184,65 @@ graph(%x: Tensor, %y: Tensor):
   }
 }
 
-void testIValueKWargs() {
-  const auto text = R"(
-    def foo(a : int, b : int, c : int = 4):
-      return a + 2*b + 3*c
-  )";
-  auto cu = compile(text);
-  auto result = cu->get_function("foo")({1}, {{"b", 3}});
-  ASSERT_EQ(result.toInt(), 19);
-}
-
-void testTemplatedOperatorCreator() {
-  constexpr char op_list[] = "foofoo::bar.template;foo::another";
+static constexpr char op_list[] = "foofoo::bar.template;foo::another";
 #define TORCH_SELECTIVE_NAME_IN_SCHEMA(l, n)                                   \
   torch::detail::SelectiveStr<c10::impl::op_whitelist_contains_name_in_schema( \
       l, n)>(n)
 
-  {
-    // Try to register an op name that does not exist in op_list.
-    // Expected: the op name is not registered.
-    torch::jit::RegisterOperators reg({OperatorGenerator(
-        TORCH_SELECTIVE_NAME_IN_SCHEMA(
-            op_list, "foofoo::not_exist(float a, Tensor b) -> Tensor"),
-        [](Stack* stack) {
-          double a;
-          at::Tensor b;
-          pop(stack, a, b);
-          push(stack, a + b);
-        },
-        aliasAnalysisFromSchema())});
+TEST(TestCustomOperator, OperatorGeneratorUndeclared) {
+  // Try to register an op name that does not exist in op_list.
+  // Expected: the op name is not registered.
+  torch::jit::RegisterOperators reg({OperatorGenerator(
+      TORCH_SELECTIVE_NAME_IN_SCHEMA(
+          op_list, "foofoo::not_exist(float a, Tensor b) -> Tensor"),
+      [](Stack* stack) {
+        double a;
+        at::Tensor b;
+        pop(stack, a, b);
+        push(stack, a + b);
+      },
+      aliasAnalysisFromSchema())});
 
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foofoo::not_exist"));
-    ASSERT_EQ(ops.size(), 0);
-  }
+  auto& ops = getAllOperatorsFor(Symbol::fromQualString("foofoo::not_exist"));
+  ASSERT_EQ(ops.size(), 0);
+}
 
-  {
-    // The operator should be successfully registered since its name is in the
-    // whitelist.
-    torch::jit::RegisterOperators reg({OperatorGenerator(
-        TORCH_SELECTIVE_NAME_IN_SCHEMA(
-            op_list, "foofoo::bar.template(float a, Tensor b) -> Tensor"),
-        [](Stack* stack) {
-          double a;
-          at::Tensor b;
-          pop(stack, a, b);
-          push(stack, a + b);
-        },
-        aliasAnalysisFromSchema())});
+TEST(TestCustomOperator, OperatorGeneratorBasic) {
+  // The operator should be successfully registered since its name is in the
+  // whitelist.
+  torch::jit::RegisterOperators reg({OperatorGenerator(
+      TORCH_SELECTIVE_NAME_IN_SCHEMA(
+          op_list, "foofoo::bar.template(float a, Tensor b) -> Tensor"),
+      [](Stack* stack) {
+        double a;
+        at::Tensor b;
+        pop(stack, a, b);
+        push(stack, a + b);
+      },
+      aliasAnalysisFromSchema())});
 
-    auto& ops = getAllOperatorsFor(Symbol::fromQualString("foofoo::bar"));
-    ASSERT_EQ(ops.size(), 1);
+  auto& ops = getAllOperatorsFor(Symbol::fromQualString("foofoo::bar"));
+  ASSERT_EQ(ops.size(), 1);
 
-    auto& op = ops.front();
-    ASSERT_EQ(op->schema().name(), "foofoo::bar");
+  auto& op = ops.front();
+  ASSERT_EQ(op->schema().name(), "foofoo::bar");
 
-    ASSERT_EQ(op->schema().arguments().size(), 2);
-    ASSERT_EQ(op->schema().arguments()[0].name(), "a");
-    ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
-    ASSERT_EQ(op->schema().arguments()[1].name(), "b");
-    ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
+  ASSERT_EQ(op->schema().arguments().size(), 2);
+  ASSERT_EQ(op->schema().arguments()[0].name(), "a");
+  ASSERT_EQ(op->schema().arguments()[0].type()->kind(), TypeKind::FloatType);
+  ASSERT_EQ(op->schema().arguments()[1].name(), "b");
+  ASSERT_EQ(op->schema().arguments()[1].type()->kind(), TypeKind::TensorType);
 
-    ASSERT_EQ(op->schema().returns().size(), 1);
-    ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
+  ASSERT_EQ(op->schema().returns().size(), 1);
+  ASSERT_EQ(op->schema().returns()[0].type()->kind(), TypeKind::TensorType);
 
-    Stack stack;
-    push(stack, 2.0f, at::ones(5));
-    op->getOperation()(&stack);
-    at::Tensor output;
-    pop(stack, output);
+  Stack stack;
+  push(stack, 2.0f, at::ones(5));
+  op->getOperation()(&stack);
+  at::Tensor output;
+  pop(stack, output);
 
-    ASSERT_TRUE(output.allclose(at::full(5, 3.0f)));
-  }
+  ASSERT_TRUE(output.allclose(at::full(5, 3.0f)));
 }
 
 } // namespace jit
