+++ test/cpp/jit/test_fuser.cpp
@@ -1,4 +1,4 @@
-#include "test/cpp/jit/test_base.h"
+#include <gtest/gtest.h>
 
 #include <torch/csrc/jit/passes/canonicalize.h>
 #include "ATen/core/interned_strings.h"
@@ -56,28 +56,27 @@
 namespace torch {
 namespace jit {
 
-void testFusion() {
-  auto testSimple = [&] {
-    const auto graph_string = R"IR(
+TEST(FuserTest, TestSimple_CUDA) {
+  const auto graph_string = R"IR(
       graph(%0 : Tensor,
             %1 : Tensor):
         %2 : Tensor = aten::mul(%0, %1)
         return (%2))IR";
-    Graph graph;
-    torch::jit::parseIR(graph_string, &graph);
+  Graph graph;
+  torch::jit::parseIR(graph_string, &graph);
 
-    auto a = at::rand({3, 4}, at::kCUDA);
-    auto b = at::rand({4, 3}, at::kCUDA).transpose(0, 1);
-    auto o = at::zeros({3, 4}, at::kCUDA);
-    auto outputs = debugLaunchGraph(graph, {a, b});
-    ASSERT_EQ(outputs.size(), 1);
-    auto o2 = a * b;
-    float max_diff = (o2 - outputs[0]).abs().max().item<double>();
-    // std::cout << "max diff: " << max_diff << "\n";
-    ASSERT_EQ(max_diff, 0);
-  };
-  testSimple();
+  auto a = at::rand({3, 4}, at::kCUDA);
+  auto b = at::rand({4, 3}, at::kCUDA).transpose(0, 1);
+  auto o = at::zeros({3, 4}, at::kCUDA);
+  auto outputs = debugLaunchGraph(graph, {a, b});
+  ASSERT_EQ(outputs.size(), 1);
+  auto o2 = a * b;
+  float max_diff = (o2 - outputs[0]).abs().max().item<double>();
+  // std::cout << "max diff: " << max_diff << "\n";
+  ASSERT_EQ(max_diff, 0);
+}
 
+TEST(FuserTest, TestOne_CUDA) {
   auto testOne = [&](int ti, int tj) {
     const auto graph_string = R"IR(
       graph(%0 : Tensor,
@@ -132,7 +131,9 @@ void testFusion() {
   testOne(0, 1);
   testOne(1, 2);
   testOne(0, 2);
+}
 
+TEST(FuserTest, FusedConcat_CUDA) {
   const auto graph_string0 = R"IR(
     graph(%0 : Tensor,
           %1 : Tensor):
@@ -175,7 +176,7 @@ void testFusion() {
   };
 }
 
-void testFusionAliasing() {
+TEST(FuserTest, FusionAliasing) {
   const auto graph_string = R"IR(
     graph(%0 : Tensor,
           %1 : Tensor):
@@ -200,7 +201,7 @@ void testFusionAliasing() {
       ->run(*g);
 }
 
-void testRegisterFusionCachesKernel() {
+TEST(FuserTest, KernelCaching) {
   // Constructs two functionally equivalent graphs
   const auto graph0_string = R"IR(
     graph(%0 : Float(2, 3, 4),
