+++ test/cpp/jit/test_misc.cpp
@@ -2225,5 +2225,15 @@ void testProfilerDisableInCallback() {
   t.join();
 }
 
+void testIValueKWargs() {
+  const auto text = R"(
+    def foo(a : int, b : int, c : int = 4):
+      return a + 2*b + 3*c
+  )";
+  auto cu = compile(text);
+  auto result = cu->get_function("foo")({1}, {{"b", 3}});
+  ASSERT_EQ(result.toInt(), 19);
+}
+
 } // namespace jit
 } // namespace torch
