+++ test/cpp/jit/tests.h
@@ -9,38 +9,26 @@
 namespace torch {
 namespace jit {
 #define TH_FORALL_TESTS(_)                        \
-  _(ADFormulas)                                   \
   _(Attributes)                                   \
   _(Blocks)                                       \
   _(CallStack)                                    \
   _(CallStackCaching)                             \
-  _(CodeTemplate)                                 \
   _(ControlFlow)                                  \
-  _(CreateAutodiffSubgraphs)                      \
-  _(CustomOperators)                              \
-  _(CustomOperatorAliasing)                       \
-  _(TemplatedOperatorCreator)                     \
   _(IValueKWargs)                                 \
   _(CustomFusion)                                 \
   _(SchemaMatching)                               \
-  _(Differentiate)                                \
-  _(DifferentiateWithRequiresGrad)                \
   _(FromQualString)                               \
   _(InternedStrings)                              \
   _(PassManagement)                               \
   _(Proto)                                        \
-  _(RegisterFusionCachesKernel)                   \
   _(SchemaParser)                                 \
   _(TopologicalIndex)                             \
   _(SubgraphUtils)                                \
   _(SubgraphUtilsVmap)                            \
   _(IRParser)                                     \
-  _(ConstantPooling)                              \
-  _(CleanUpPasses)                                \
   _(THNNConv)                                     \
   _(ATenNativeBatchNorm)                          \
   _(NoneSchemaMatch)                              \
-  _(ClassParser)                                  \
   _(UnifyTypes)                                   \
   _(Profiler)                                     \
   _(FallbackGraphs)                               \
@@ -61,15 +49,10 @@ namespace jit {
   _(ModuleDeepcopyAliasing)                       \
   _(ModuleDefine)                                 \
   _(QualifiedName)                                \
-  _(ClassImport)                                  \
-  _(ScriptObject)                                 \
   _(ExtraFilesHookPreference)                     \
   _(SaveExtraFilesHook)                           \
   _(TypeTags)                                     \
-  _(DCE)                                          \
   _(CustomFusionNestedBlocks)                     \
-  _(ClassDerive)                                  \
-  _(SaveLoadTorchbind)                            \
   _(ModuleInterfaceSerialization)                 \
   _(ModuleCloneWithModuleInterface)               \
   _(ClassTypeAddRemoveAttr)                       \
@@ -100,7 +83,6 @@ namespace jit {
   _(LiteInterpreterHierarchyModuleInfo)           \
   _(LiteInterpreterDuplicatedClassTypeModuleInfo) \
   _(LiteInterpreterEval)                          \
-  _(TorchbindIValueAPI)                           \
   _(LiteInterpreterDict)                          \
   _(LiteInterpreterFindAndRunMethod)              \
   _(LiteInterpreterFindWrongMethodName)           \
@@ -109,12 +91,10 @@ namespace jit {
   _(MobileSaveLoadParameters)                     \
   _(MobileSaveLoadParametersEmpty)                \
   _(LiteSGD)                                      \
-  _(LiteSequentialSampler)                        \
-  _(FusionAliasing)
+  _(LiteSequentialSampler)
 
 #if defined(USE_CUDA)
 #define TH_FORALL_TESTS_CUDA(_)                     \
-  _(Fusion)                                         \
   _(GraphExecutor)                                  \
   _(ModuleConversion)                               \
   _(Interp)                                         \
@@ -219,7 +199,6 @@ namespace jit {
   _(GPU_FusionThreadPredicate)
 #else
 #define TH_FORALL_TESTS_CUDA(_) \
-  _(Fusion)                     \
   _(GraphExecutor)              \
   _(ModuleConversion)           \
   _(Interp)                     \
